backend/requirements.txt
fastapi
uvicorn[standard]

langchain
langchain-community
langchain-core
langchain-text-splitters

sentence-transformers
faiss-cpu

pydantic
python-dotenv

pypdf
PyPDF2
python-docx

google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

groq
langchain-groq